Add an autospliter for Hexcells
I made this to learn about ptrace and get accurate splits while speedrunning
hexcells. It has been long enought I don't remember the exact details, but I
think there is still a bug with it. However, I want to get it tracked so that I
don't lose it.

Additionally there is a rust based debugger. Not sure if I will commit that yet.